backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
djangorestframework-simplejwt>=5.2.0
django-filter>=23.0
django-cors-headers>=4.0
python-decouple>=3.8
dj-database-url>=2.1.0
